Move item creation logic to its own class
